+++ data/CulturalClubCrew.ts
@@ -1,0 +1,283 @@
+import { StaticImageData } from "next/image";
+
+interface CrewProps {
+    img: StaticImageData;
+    name: string;
+    post: string;
+}
+
+// import crew
+import Panshul_Senapati from "../assets/team_23-24/Panshul_Senapati.jpg";
+import Aseem_Rai from "../assets/team_23-24/Aseem_Rai.jpg";
+import Ashutosh_Jaiswal from "../assets/team_23-24/Ashutosh_Jaiswal.jpeg";
+import Arman_Singh_Kshatri from "../assets/team_23-24/Male-member.jpeg";
+import Ojash_Verma from "../assets/team_23-24/Male-member.jpeg";
+import Simmi_Baghel from "../assets/team_23-24/Simmi_Baghel.jpg";
+import Priyanshu_Srivastava from "../assets/team_23-24/Priyanshu_Srivastava.jpg";
+import Anoushka_Koul from "../assets/team_23-24/Anoushka_koul.jpeg";
+import Tanisi_Jha from "../assets/team_23-24/Female-member.jpeg";
+import Varsha_Sri_Narla from "../assets/team_23-24/Varsha_Sri_Narla.jpg";
+import Akhilesh_Vajrala from "../assets/team_23-24/Male-member.jpeg";
+import Mohd_Kaif from "../assets/team_23-24/Male-member.jpeg";
+import Swastika_Satya from "../assets/team_23-24/Swastika_Satya.jpg";
+import Vishal_Sharma from "../assets/team_23-24/Vishal_Sharma.jpg";
+import Kanika_Malhotra from "../assets/team_23-24/Kanika_Malhotra -.jpg";
+import Juttuka_Saaketh from "../assets/team_23-24/Male-member.jpeg";
+import Anshu_kashyap from "../assets/team_23-24/Female-member.jpeg";
+import Krishna_aggarwal from "../assets/team_23-24/Krishna_Agrawal.jpg";
+import Niharika_Murmu from "../assets/team_23-24/Niharika_Murmu.jpg";
+import Satvik_Yadav from "../assets/team_23-24/Satvik_Yadav.jpg";
+import RamaR_Katta from "../assets/team_23-24/Male-member.jpeg";
+import Aryan_Kumar_Baghel from "../assets/team_23-24/Aryan_Kumar_Baghel.jpg";
+import Anuja_Singh from "../assets/team_23-24/Female-member.jpeg";
+import Eluri_Rithwik from "../assets/team_23-24/Eluri_Rithwik.jpg";
+import Muskan_Dewangan from "../assets/team_23-24/Muskan_Dewangan.jpg";
+import Thanisha_Dewangan from "../assets/team_23-24/Thanisha _Dewangan..jpg";
+import Shalini_Singh from "../assets/team_23-24/Shalini_Singh.jpg";
+import Sanjana_Sori from "../assets/team_23-24/Sanjana_Sori.jpg";
+import Vajrala_Akhilesh from "../assets/team_23-24/Male-member.jpeg";
+import P_Punit from "../assets/team_23-24/Male-member.jpeg";
+import Anjali_Daheriya from "../assets/team_23-24/Female-member.jpeg";
+import Aniket_Pandey from "../assets/team_23-24/Aniket_Pandey.jpg";
+import Shreya_Tiwari from "../assets/team_23-24/Shreya_Tiwari.jpg";
+import Ayush_Yadav from "../assets/team_23-24/Ayush_Yadav.jpg";
+import Muskan_dewangan from "../assets/team_23-24/Muskan_Dewangan.jpg";
+import Apurva_Sahu from "../assets/team_23-24/Apurva_Sahu.jpg";
+import Aayush_Kulkarni from "../assets/team_23-24/Aayush_Kulkarni.jpg";
+import Anoushka_Kaul from "../assets/team_23-24/Anoushka_koul.jpeg";
+import Trisul from "../assets/team_23-24/Trisul.jpeg";
+import Shaurya_Malhan from "../assets/team_23-24/Shaurya_Malhan.jpeg";
+
+export const ytc: CrewProps[] = []
+export const tdb: CrewProps[] = [
+    {
+        img: Shreya_Tiwari,
+        name: "Shreya Tiwari",
+        post: "Club Secretary",
+    },
+    {
+        img: Aayush_Kulkarni,
+        name: "Aayush Kulkarni",
+        post: "Vice Secretary",
+    },
+    {
+        img: Anoushka_Kaul,
+        name: "Anoushka Kaul",
+        post: "Core Member",
+    },
+    {
+        img: Trisul,
+        name: "Trisul",
+        post: "Core Member",
+    },
+    {
+        img: Shaurya_Malhan,
+        name: "Shaurya Malhan",
+
+        post: "Core Member",
+    }, {
+        img: Aniket_Pandey,
+        name: "Aniket Pandey",
+        post: "Core Member",
+    },
+];
+export const mun: CrewProps[] = [
+    {
+        img: Kanika_Malhotra,
+        name: "Kanika Malhotra",
+        post: "Club Secretary",
+    },
+    {
+        img: Aniket_Pandey,
+        name: "Aniket Pandey",
+        post: "Vice Secretary",
+    },
+    {
+        img: Shreya_Tiwari,
+        name: "Shreya Tiwari",
+        post: "Core Member",
+    },
+    {
+        img: Ayush_Yadav,
+        name: "Ayush Yadav",
+        post: "Core Member",
+    },
+    {
+        img: Muskan_dewangan,
+        name: "Muskan Dewangan",
+        post: "Core Member",
+    },
+    {
+        img: Apurva_Sahu,
+        name: "Apurva Sahu",
+        post: "Core Member",
+    },
+];
+export const indradhanush: CrewProps[] = [
+    {
+        img: Thanisha_Dewangan,
+        name: "Thanisha Dewangan",
+        post: "Club Secretary",
+    },
+    {
+        img: Shalini_Singh,
+        name: "Shalini Singh",
+        post: "Vice Secretary",
+    },
+    {
+        img: Sanjana_Sori,
+        name: "Sanjana Sori",
+        post: "Core Member",
+    },
+    {
+        img: Vajrala_Akhilesh,
+        name: "Vajrala Akhilesh",
+        post: "Core Member",
+    },
+    {
+        img: P_Punit,
+        name: "P Punit",
+        post: "Core Member",
+    },
+    {
+        img: Anjali_Daheriya,
+        name: "Anjali Daheriya",
+        post: "Core Member",
+    },
+];
+
+export const igniters: CrewProps[] = [
+    {
+        img: Niharika_Murmu,
+        name: "Niharika Murmu",
+        post: "Club Secretary",
+    },
+    {
+        img: Satvik_Yadav,
+        name: "Satvik Yadav",
+        post: "Co-Vice Secretary",
+    },
+    {
+        img: RamaR_Katta,
+        name: "Rama Rakshith Katta",
+        post: "Co-Vice Secretary",
+    },
+    {
+        img: Aryan_Kumar_Baghel,
+        name: "Aryan Kumar Baghel",
+        post: "core member",
+    },
+    {
+        img: Anuja_Singh,
+        name: "Anuja Singh Pandaw",
+        post: "core member",
+    },
+    {
+        img: Eluri_Rithwik,
+        name: "Elluri Rithwik",
+        post: "core member",
+    },
+    {
+        img: Muskan_Dewangan,
+        name: "Muskan Dewangan",
+        post: "core member",
+    },
+];
+export const cdt: CrewProps[] = [
+    {
+        img: Swastika_Satya,
+        name: "Swastika Satya",
+        post: "Club Secretary",
+    },
+    {
+        img: Vishal_Sharma,
+        name: "Vishal Sharma",
+        post: "Vice Secretary",
+    },
+    {
+        img: Kanika_Malhotra,
+        name: "Kanika Malhotra",
+        post: "Core member",
+    },
+    {
+        img: Juttuka_Saaketh,
+        name: "Juttuka Saaketh",
+        post: "Core Member",
+    },
+    {
+        img: Anshu_kashyap,
+        name: "Anhsu Kashyap",
+        post: "Core Member",
+    },
+    {
+        img: Krishna_aggarwal,
+        name: "Krishna Aggarwal",
+        post: "Core Member",
+    },
+
+];
+
+export const capriccio: CrewProps[] = [
+    {
+        img: Priyanshu_Srivastava,
+        name: "Priyanshu Srivastava",
+        post: "Club Secretary",
+    },
+    {
+        img: Anoushka_Koul,
+        name: "Anoushka Koul",
+        post: "Vice Secretary",
+    },
+    {
+        img: Tanisi_Jha,
+        name: "Tanisi Jha",
+        post: "Core Member",
+    },
+    {
+        img: Varsha_Sri_Narla,
+        name: "Varsha Sri Narla",
+        post: "Core Member",
+    },
+    {
+        img: Akhilesh_Vajrala,
+        name: "Akhilesh Vajrala",
+        post: "Core Member",
+    },
+    {
+        img: Mohd_Kaif,
+        name: "Mohd Kaif",
+        post: "Core Member",
+    },
+];
+export const c2c: CrewProps[] = [
+    {
+        img: Panshul_Senapati,
+        name: "Panshul Senapati",
+        post: "Club Secretary",
+    },
+    {
+        img: Aseem_Rai,
+        name: "Aseem Rai",
+        post: "Vice Secretary",
+    },
+    {
+        img: Ashutosh_Jaiswal,
+        name: "Ashutosh Jaiswal",
+        post: "Core Member",
+    },
+    {
+        img: Ojash_Verma,
+        name: "Ojash Verma",
+        post: "Core Member",
+    },
+    {
+        img: Arman_Singh_Kshatri,
+        name: "Arman Singh Kshatri",
+        post: "Core Member",
+    },
+    {
+        img: Simmi_Baghel,
+        name: "Simmi Baghel",
+        post: "Core Member",
+    },
+];
