+++ data/MediaClubCrew.ts
@@ -1,0 +1,162 @@
+import { StaticImageData } from "next/image";
+
+interface CrewProps {
+    img: StaticImageData;
+    name: string;
+    post: string;
+}
+
+// import crew
+import Ojash_Verma from "../assets/team_23-24/Male-member.jpeg";
+import Aayush_Kulkarni from "../assets/team_23-24/Aayush_Kulkarni.jpg";
+import Trisul from "../assets/team_23-24/Trisul.jpeg";
+import Shreyansh_Kushwaha from "../assets/team_23-24/Shreyansh_Kushwaha.jpg"
+import Ishita_Tegar from "../assets/team_23-24/Ishita_Tegar.jpg";
+import Raj_Bhattacharya from "../assets/team_23-24/Male-member.jpeg";
+import A_Satya_Kartik from "../assets/team_23-24/Male-member.jpeg";
+import S_Chandu from "../assets/team_23-24/Male-member.jpeg";
+import Tanishka from "../assets/team_23-24/Female-member.jpeg";
+import Aryansh_Verma from "../assets/team_23-24/Aryansh_Verma.jpg";
+import Abhijeet_Jharbade from "../assets/team_23-24/Male-member.jpeg";
+import ACHANTA_SATYA_KARTHIK from "../assets/team_23-24/Satya_Karthik_Achanta.jpg";
+import ANKITAN_KUMAR from "../assets/team_23-24/Ankitan_Kumar.jpg";
+import ANUJA_SINGH_PANDAW from "../assets/team_23-24/Female-member.jpeg";
+import APURVA_SAHU from "../assets/team_23-24/Apurva_Sahu.jpg";
+import KARTHIKEYA_PRACHODRAN from "../assets/team_23-24/Male-member.jpeg";
+import RIMJHIM_SHARMA from "../assets/team_23-24/Female-member.jpeg";
+import AAYUSH_KULKARNI from "../assets/team_23-24/Aayush_Kulkarni.jpg";
+import ARYAN_KUMAR_BAGHEL from "../assets/team_23-24/Aryan_Kumar_Baghel.jpg";
+import Manu_Sai from "../assets/team_23-24/Manu_Sai.jpg";
+import Ojas_Dubey1 from "../assets/team_23-24/Ojas_Dubey.jpg";
+import Yashraj from "../assets/team_23-24/Yash_Raj.jpeg";
+
+export const website: CrewProps[] = [
+    {
+        img: Ojas_Dubey1,
+        name: "Ojas Dubey",
+        post: "Head",
+    },
+    {
+        img: Manu_Sai,
+        name: "Manu Sai",
+        post: "Core Member",
+    },
+    {
+        img: Shreyansh_Kushwaha,
+        name: "Shreyansh Kushwaha",
+        post: "Core Member",
+    },
+    {
+        img: Yashraj,
+        name: "Yashraj",
+        post: "Core Member",
+    },
+];
+
+export const shutterbug: CrewProps[] = [
+    {
+        img: Abhijeet_Jharbade,
+        name: "Abhijeet Jharbade",
+        post: "Head",
+    },
+    {
+        img: ACHANTA_SATYA_KARTHIK,
+        name: "Achanta Satya Karthik",
+        post: "Core member",
+    },
+    {
+        img: ANKITAN_KUMAR,
+        name: "Ankitan Kumar",
+        post: "Core member",
+    },
+    {
+        img: ANUJA_SINGH_PANDAW,
+        name: "Anuja Singh Pandaw",
+        post: "Core member",
+    },
+    {
+        img: APURVA_SAHU,
+        name: "Apurva Sahu",
+        post: "Core member",
+    },
+    {
+        img: KARTHIKEYA_PRACHODRAN,
+        name: "Karthikeya Prachodran",
+        post: "Core member",
+    },
+    {
+        img: RIMJHIM_SHARMA,
+        name: "Rimjhim Sharma",
+        post: "Core member",
+    },
+    {
+        img: AAYUSH_KULKARNI,
+        name: "Aayush Kulkarni",
+        post: "Core member",
+    },
+    {
+        img: ARYAN_KUMAR_BAGHEL,
+        name: "Aryan Kumar Baghel",
+        post: "Core member",
+    },
+];
+export const pr: CrewProps[] = [
+    {
+        img: Aryansh_Verma,
+        name: "Aryansh Verma",
+        post: "Head",
+    },
+    {
+        img: Aayush_Kulkarni,
+        name: "Aayush Kulkarni",
+        post: "Core member",
+    },
+    {
+        img: Raj_Bhattacharya,
+        name: "Raj Bhattacharya",
+        post: "Core member",
+    },
+    {
+        img: Ishita_Tegar,
+        name: "Ishita Tegar",
+        post: "Core member",
+    },
+    {
+        img: Trisul,
+        name: "Trishul",
+        post: "Core member",
+    },
+];
+
+export const design: CrewProps[] = [
+    {
+        img: A_Satya_Kartik,
+        name: "A Satya Karthik",
+        post: "Head",
+    },
+    {
+        img: S_Chandu,
+        name: "S.Chandu",
+        post: "Core Member",
+    },
+    {
+        img: Ojash_Verma,
+        name: "Ojash Verma",
+        post: "Core Member",
+    },
+    {
+        img: Tanishka,
+        name: "Tanishka Chauhan",
+        post: "Core Member",
+    },
+    {
+        img: Shreyansh_Kushwaha,
+        name: "Shreyansh Kushwaha",
+        post: "Core Member",
+    },
+    {
+        img: Trisul,
+        name: "Trishul",
+        post: "Core Member",
+    },
+];
