+++ data/TeamData.tsx
@@ -1,102 +1,102 @@
 import { StaticImageData } from "next/image";
 
 // 2023-2024
-import Aditya_Oza from "../../assets/team_23-24/Aditya_Oza.jpg";
-import Kanika from "../../assets/team_23-24/Kanika_Chaurasia.jpg";
-import Shivam_Kushwaha from "../../assets/team_23-24/Shivam_Kushwaha.jpg";
-import Ch_Swati from "../../assets/team_23-24/Ch_Swati.jpg";
-import Akshat_Shree_Mishra from "../../assets/team_23-24/Akshat_Shree_Mishra.jpg";
-import Rimjhim_Sharma from "../../assets/team_23-24/Rimjhim_Sharma.jpg";
-import Deependra_Singh_Bhow from "../../assets/team_23-24/Deependar.jpeg";
-import Kusum_Patel from "../../assets/team_23-24/Kusum_Patel.jpg";
-import Jay_Padia from "../../assets/team_23-24/Jay_Padia.jpg";
-import R_Sai_Shivani from "../../assets/team_23-24/Sai_Shivani.jpg";
-import Kamal from "../../assets/team_23-24/kamal.jpeg";
+import Aditya_Oza from "../assets/team_23-24/Aditya_Oza.jpg";
+import Kanika from "../assets/team_23-24/Kanika_Chaurasia.jpg";
+import Shivam_Kushwaha from "../assets/team_23-24/Shivam_Kushwaha.jpg";
+import Ch_Swati from "../assets/team_23-24/Ch_Swati.jpg";
+import Akshat_Shree_Mishra from "../assets/team_23-24/Akshat_Shree_Mishra.jpg";
+import Rimjhim_Sharma from "../assets/team_23-24/Rimjhim_Sharma.jpg";
+import Deependra_Singh_Bhow from "../assets/team_23-24/Deependar.jpeg";
+import Kusum_Patel from "../assets/team_23-24/Kusum_Patel.jpg";
+import Jay_Padia from "../assets/team_23-24/Jay_Padia.jpg";
+import R_Sai_Shivani from "../assets/team_23-24/Sai_Shivani.jpg";
+import Kamal from "../assets/team_23-24/kamal.jpeg";
 
 // 2022-2023
-import SristhiSahooUG from "../../assets/team_22-23/SristhiSahooUG.jpeg";
-import HimanshuYadav_UG_2022 from "../../assets/team_22-23/HimanshuYadav_UG_2022.jpeg";
-import Nishok from "../../assets/team_22-23/Nishok.jpg";
-import Supriya from "../../assets/team_22-23/Supriya.jpeg";
-import Aaditya_Kumar from "../../assets/team_22-23/Aaditya_Kumar.jpeg";
-import SonaliB from "../../assets/team_22-23/SonaliB.jpeg";
-import Kishan_Mishra from "../../assets/team_22-23/Kishan_Mishra.jpeg";
-import Nidhi_Vaishnav from "../../assets/team_22-23/Nidhi_Vaishnav.jpeg";
-import Soumya_min from "../../assets/team_22-23/Soumya-min.jpeg";
-import Sudhanshu_Tripathi from "../../assets/team_22-23/Sudhanshu_Tripathi.jpeg";
-import Deepesh_Agrawal from "../../assets/team_22-23/Deepesh_Agrawal.jpeg";
-import Ojas_Dubey from "../../assets/team_22-23/Ojas_Dubey.jpeg";
-import Aman_Singh from "../../assets/team_22-23/Aman_Singh.jpeg";
-import Vinayak_Shrivastava from "../../assets/team_22-23/Vinayak_Shrivastava.jpeg";
-import saiprabhat from "../../assets/team_22-23/saiprabhat.jpeg";
-import Shubham_khilari from "../../assets/team_22-23/Shubham_khilari.jpeg";
-import nidhibhaskar from "../../assets/team_22-23/nidhibhaskar.jpeg";
-import navneet_chaubey from "../../assets/team_22-23/navneet_chaubey.jpeg";
-import Jay_Deep_Singh from "../../assets/team_22-23/Jay_Deep_Singh.jpeg";
-import manoj_SacP from "../../assets/team_22-23/manoj.jpeg";
-import Karan_Agrawal from "../../assets/team_22-23/Karan_Agrawal.jpeg";
+import SristhiSahooUG from "../assets/team_22-23/SristhiSahooUG.jpeg";
+import HimanshuYadav_UG_2022 from "../assets/team_22-23/HimanshuYadav_UG_2022.jpeg";
+import Nishok from "../assets/team_22-23/Nishok.jpg";
+import Supriya from "../assets/team_22-23/Supriya.jpeg";
+import Aaditya_Kumar from "../assets/team_22-23/Aaditya_Kumar.jpeg";
+import SonaliB from "../assets/team_22-23/SonaliB.jpeg";
+import Kishan_Mishra from "../assets/team_22-23/Kishan_Mishra.jpeg";
+import Nidhi_Vaishnav from "../assets/team_22-23/Nidhi_Vaishnav.jpeg";
+import Soumya_min from "../assets/team_22-23/Soumya-min.jpeg";
+import Sudhanshu_Tripathi from "../assets/team_22-23/Sudhanshu_Tripathi.jpeg";
+import Deepesh_Agrawal from "../assets/team_22-23/Deepesh_Agrawal.jpeg";
+import Ojas_Dubey from "../assets/team_22-23/Ojas_Dubey.jpeg";
+import Aman_Singh from "../assets/team_22-23/Aman_Singh.jpeg";
+import Vinayak_Shrivastava from "../assets/team_22-23/Vinayak_Shrivastava.jpeg";
+import saiprabhat from "../assets/team_22-23/saiprabhat.jpeg";
+import Shubham_khilari from "../assets/team_22-23/Shubham_khilari.jpeg";
+import nidhibhaskar from "../assets/team_22-23/nidhibhaskar.jpeg";
+import navneet_chaubey from "../assets/team_22-23/navneet_chaubey.jpeg";
+import Jay_Deep_Singh from "../assets/team_22-23/Jay_Deep_Singh.jpeg";
+import manoj_SacP from "../assets/team_22-23/manoj.jpeg";
+import Karan_Agrawal from "../assets/team_22-23/Karan_Agrawal.jpeg";
 
 // 2021-2022
-import Dr_Venkanna from "../../assets/team_21-22/venkanna_u.jpg";
-import ketan from "../../assets/team_21-22/ketan.jpeg";
-import abheeshth from "../../assets/team_21-22/abheeshth.jpg";
-import Sakshi_Verma from "../../assets/team_21-22/Sakshi Verma.jpg";
-import saumya from "../../assets/team_21-22/saumya .jpg";
-import ayush_tirkey from "../../assets/team_20-21/ayush_tirkey.jpg";
-import Meetiksha from "../../assets/team_21-22/Meetiksha.jpeg";
-import gunateja from "../../assets/team_21-22/gunateja.jpg";
-import Roopa_navya from "../../assets/team_21-22/Roopa navya.jpg";
-import shashank from "../../assets/team_21-22/shashank.jpg";
-import Suraj from "../../assets/team_21-22/Suraj.jpeg";
-import Anupam from "../../assets/team_21-22/Anupam.jpg";
-import amogh from "../../assets/team_21-22/amogh.jpg";
-import Himanshu from "../../assets/team_21-22/Himanshu.jpeg";
-import Soumya from "../../assets/team_21-22/Soumya.jpg";
-import Tavisha from "../../assets/team_21-22/Tavisha.jpg";
-import Sonal_Singh from "../../assets/team_21-22/PSX_20200229_183102.jpg";
-import Shubham from "../../assets/team_21-22/Shubham.jpg";
-import Bhupendra_Chouhan from "../../assets/team_21-22/Bhupendra Chouhan_.jpg";
+import Dr_Venkanna from "../assets/team_21-22/venkanna_u.jpg";
+import ketan from "../assets/team_21-22/ketan.jpeg";
+import abheeshth from "../assets/team_21-22/abheeshth.jpg";
+import Sakshi_Verma from "../assets/team_21-22/Sakshi Verma.jpg";
+import saumya from "../assets/team_21-22/saumya .jpg";
+import ayush_tirkey from "../assets/team_20-21/ayush_tirkey.jpg";
+import Meetiksha from "../assets/team_21-22/Meetiksha.jpeg";
+import gunateja from "../assets/team_21-22/gunateja.jpg";
+import Roopa_navya from "../assets/team_21-22/Roopa navya.jpg";
+import shashank from "../assets/team_21-22/shashank.jpg";
+import Suraj from "../assets/team_21-22/Suraj.jpeg";
+import Anupam from "../assets/team_21-22/Anupam.jpg";
+import amogh from "../assets/team_21-22/amogh.jpg";
+import Himanshu from "../assets/team_21-22/Himanshu.jpeg";
+import Soumya from "../assets/team_21-22/Soumya.jpg";
+import Tavisha from "../assets/team_21-22/Tavisha.jpg";
+import Sonal_Singh from "../assets/team_21-22/PSX_20200229_183102.jpg";
+import Shubham from "../assets/team_21-22/Shubham.jpg";
+import Bhupendra_Chouhan from "../assets/team_21-22/Bhupendra Chouhan_.jpg";
 
 // 2020-2021
-import prince_yadav from "../../assets/team_20-21/prince_yadav.jpg";
-import Vidhi_Sharma from "../../assets/team_20-21/vidhi_sharma.jpg";
-import alluri_lsv_siddhartha_varma from "../../assets/team_20-21/alluri_lsv_siddhartha_varma.jpg";
-import sakshi_kumari from "../../assets/team_20-21/sakshi_kumari.jpg";
-import samar_lal from "../../assets/team_20-21/samar_lal.jpg";
-import palak_chandrakar from "../../assets/team_20-21/palak_chandrakar.jpg";
-import abhay_bagh from "../../assets/team_20-21/abhay_bagh.jpg";
-import rashi_bagga from "../../assets/team_20-21/rashi_bagga.jpeg";
-import saumya_bhatnagar from "../../assets/team_20-21/saumya_bhatnagar.jpeg";
-import himanshu_patel from "../../assets/team_20-21/himanshu_patel.jpeg";
-import siddharth_singh from "../../assets/team_20-21/siddharth_singh.jpg";
-import vipul_sangode from "../../assets/team_20-21/vipul_sangode.jpeg";
-import ankit_sharma from "../../assets/team_20-21/ankit_sharma.jpeg";
-import annapragada_nishant from "../../assets/team_20-21/annapragada_nishant.jpg";
-import shreyas_lall from "../../assets/team_20-21/shreyas_lall.jpeg";
-import apurv_rai from "../../assets/team_20-21/apurv_rai.jpeg";
-import shalini_lodhi from "../../assets/team_20-21/shalini_lodhi.jpg";
-import shivangini_prasad from "../../assets/team_20-21/shivangini_prasad.jpg";
-import sonal_singh from "../../assets/team_20-21/sonal_singh.jpg";
+import prince_yadav from "../assets/team_20-21/prince_yadav.jpg";
+import Vidhi_Sharma from "../assets/team_20-21/vidhi_sharma.jpg";
+import alluri_lsv_siddhartha_varma from "../assets/team_20-21/alluri_lsv_siddhartha_varma.jpg";
+import sakshi_kumari from "../assets/team_20-21/sakshi_kumari.jpg";
+import samar_lal from "../assets/team_20-21/samar_lal.jpg";
+import palak_chandrakar from "../assets/team_20-21/palak_chandrakar.jpg";
+import abhay_bagh from "../assets/team_20-21/abhay_bagh.jpg";
+import rashi_bagga from "../assets/team_20-21/rashi_bagga.jpeg";
+import saumya_bhatnagar from "../assets/team_20-21/saumya_bhatnagar.jpeg";
+import himanshu_patel from "../assets/team_20-21/himanshu_patel.jpeg";
+import siddharth_singh from "../assets/team_20-21/siddharth_singh.jpg";
+import vipul_sangode from "../assets/team_20-21/vipul_sangode.jpeg";
+import ankit_sharma from "../assets/team_20-21/ankit_sharma.jpeg";
+import annapragada_nishant from "../assets/team_20-21/annapragada_nishant.jpg";
+import shreyas_lall from "../assets/team_20-21/shreyas_lall.jpeg";
+import apurv_rai from "../assets/team_20-21/apurv_rai.jpeg";
+import shalini_lodhi from "../assets/team_20-21/shalini_lodhi.jpg";
+import shivangini_prasad from "../assets/team_20-21/shivangini_prasad.jpg";
+import sonal_singh from "../assets/team_20-21/sonal_singh.jpg";
 
 // 2019-2020
-import nitin_chandra from "../../assets/team_19-20/nitin_chandra.jpg";
-import sriyuta_srivastava from "../../assets/team_19-20/sriyuta_srivastava.jpg";
-import jatin_aditya_reddy from "../../assets/team_19-20/jatin_aditya_reddy.jpg";
-import tina_sahu from "../../assets/team_19-20/tina_sahu.jpeg";
-import rituraj_kushwaha from "../../assets/team_19-20/rituraj_kushwaha.jpg";
-import manorama_singh_rathore from "../../assets/team_19-20/manorama_singh_rathore.jpg";
-import pushpak_singh_tawar from "../../assets/team_19-20/pushpak_singh_tawar.jpg";
-import likitha_bonda from "../../assets/team_19-20/likitha_bonda.jpg";
+import nitin_chandra from "../assets/team_19-20/nitin_chandra.jpg";
+import sriyuta_srivastava from "../assets/team_19-20/sriyuta_srivastava.jpg";
+import jatin_aditya_reddy from "../assets/team_19-20/jatin_aditya_reddy.jpg";
+import tina_sahu from "../assets/team_19-20/tina_sahu.jpeg";
+import rituraj_kushwaha from "../assets/team_19-20/rituraj_kushwaha.jpg";
+import manorama_singh_rathore from "../assets/team_19-20/manorama_singh_rathore.jpg";
+import pushpak_singh_tawar from "../assets/team_19-20/pushpak_singh_tawar.jpg";
+import likitha_bonda from "../assets/team_19-20/likitha_bonda.jpg";
 
 // 2018-2019
-import vishal_pallagani from "../../assets/team_18-19/vishal_pallagani.jpg";
-import ankita_jha from "../../assets/team_18-19/ankita_jha.jpg";
-import kushashwa_ravi_shrimali from "../../assets/team_18-19/kushashwa_ravi_shrimali.png";
-import priyanka_mall from "../../assets/team_18-19/priyanka_mall.jpg";
-import rasheed_ahmed from "../../assets/team_18-19/rasheed_ahmed.jpg";
-import aditia_agrahari from "../../assets/team_18-19/aditia_agrahari.jpg";
-import nirbhay_giri from "../../assets/team_18-19/nirbhay_giri.jpg";
-import minakshee_shukla from "../../assets/team_18-19/minakshee_shukla.jpg";
+import vishal_pallagani from "../assets/team_18-19/vishal_pallagani.jpg";
+import ankita_jha from "../assets/team_18-19/ankita_jha.jpg";
+import kushashwa_ravi_shrimali from "../assets/team_18-19/kushashwa_ravi_shrimali.png";
+import priyanka_mall from "../assets/team_18-19/priyanka_mall.jpg";
+import rasheed_ahmed from "../assets/team_18-19/rasheed_ahmed.jpg";
+import aditia_agrahari from "../assets/team_18-19/aditia_agrahari.jpg";
+import nirbhay_giri from "../assets/team_18-19/nirbhay_giri.jpg";
+import minakshee_shukla from "../assets/team_18-19/minakshee_shukla.jpg";
 
 interface CrewProps {
   img: StaticImageData;
