+++ data/TechnicalClubCrew.ts
@@ -1,0 +1,198 @@
+import { StaticImageData } from "next/image";
+
+interface CrewProps {
+    img: StaticImageData;
+    name: string;
+    post: string;
+}
+
+// import crew
+import Ashutosh_Jaiswal from "../assets/team_23-24/Ashutosh_Jaiswal.jpeg";
+import Arman_Singh_Kshatri from "../assets/team_23-24/Male-member.jpeg";
+import Swastika_Satya from "../assets/team_23-24/Swastika_Satya.jpg";
+import Vishal_Sharma from "../assets/team_23-24/Vishal_Sharma.jpg";
+import Aniket_Pandey from "../assets/team_23-24/Aniket_Pandey.jpg";
+import Ayush_Yadav from "../assets/team_23-24/Ayush_Yadav.jpg";
+import Shaurya_Malhan from "../assets/team_23-24/Shaurya_Malhan.jpeg";
+import YashV_Gautum from "../assets/team_23-24/Yash_Vardhan_Gautam.jpg";
+import Sanskar_Singh from "../assets/team_23-24/Sanskar_Singh.jpeg";
+import Arya_Mahesh from "../assets/team_23-24/Arya_Mahesh_Bhiwapurkar.jpg";
+import Vaibhav_Suntwal from "../assets/team_23-24/Vaibhav_Suntwal.jpg";
+import Shivaibhav from "../assets/team_23-24/Shivaibhav_Dewangan.jpg";
+import Gaurish from "../assets/team_23-24/Gaurish_Ojha.jpg";
+import Dhyanendra from "../assets/team_23-24/Dhyanendra_Tripathi.jpg";
+import Swasti_Srivastava from "../assets/team_23-24/Swasti_Srivastava.jpeg";
+import Manav_Raj from "../assets/team_23-24/Manav_Raj_Singh.jpg";
+import V_Jaswanth from "../assets/team_23-24/Vemula_Jashwanth.jpeg";
+import Alankrit from "../assets/team_23-24/Alankrit_Sharma.jpg"
+import Shreyansh_Kushwaha from "../assets/team_23-24/Shreyansh_Kushwaha.jpg"
+import Venkata_surya from "../assets/team_23-24/Venkata_Surya_Sundar_Vadali.jpg"
+import Priyanshu_shri from "../assets/team_23-24/Priyanshu_Srivastava.jpg"
+import Ishita_Tegar from "../assets/team_23-24/Ishita_Tegar.jpg";
+import Umesh_Sinha from "../assets/team_23-24/Umesh_Sinha.jpg";
+import Alankar_Saxena from "../assets/team_23-24/Alankar_Saxena.jpg";
+import Namita from "../assets/team_23-24/Namita_shah.jpg";
+import Krishna_Aggarwal from "../assets/team_23-24/Krishna_Agrawal.jpg";
+import Bhavik from "../assets/team_23-24/Male-member.jpeg";
+import Harsh_Shrivastava from "../assets/team_23-24/Harsh_Shrivastava.jpg";
+import Lokesh_Harmani from "../assets/team_23-24/Male-member.jpeg";
+
+export const inquizitive: CrewProps[] = [
+    {
+        img: Ayush_Yadav,
+        name: "Ayush Yadav",
+        post: "Club Secretary",
+    },
+    {
+        img: Bhavik,
+        name: "Bhavik Sachdev",
+        post: "Vice Secretary",
+    },
+    {
+        img: Aniket_Pandey,
+        name: "Aniket Pandey",
+        post: "Core Member",
+    },
+    {
+        img: Arman_Singh_Kshatri,
+        name: "Arman Singh Kshatri",
+        post: "Core Member",
+    },
+    {
+        img: Harsh_Shrivastava,
+        name: "Harsh Shrivastava",
+        post: "Core Member",
+    },
+    {
+        img: Lokesh_Harmani,
+        name: "Lokesh Hariramani",
+        post: "Core Member",
+    },
+];
+
+export const iic: CrewProps[] = []
+export const comet: CrewProps[] = [
+    {
+        img: Ishita_Tegar,
+        name: "Ishita Tegar",
+        post: "Club Secretary",
+    },
+    {
+        img: Umesh_Sinha,
+        name: "Umesh Sinha",
+        post: "Vice Secretary",
+    },
+    {
+        img: Alankar_Saxena,
+        name: "Alankar Saxena",
+        post: "Core Member",
+    },
+    {
+        img: Namita,
+        name: "Namita Shah",
+        post: "Core Member",
+    },
+    {
+        img: Krishna_Aggarwal,
+        name: "Krishna Aggarwal",
+        post: "Core Member",
+    },
+];
+export const ciphercell: CrewProps[] = [
+    {
+        img: V_Jaswanth,
+        name: "Jaswanth Vemula",
+        post: "Club Secretary",
+    },
+    {
+        img: Alankrit,
+        name: "Alankrit Sharma",
+        post: "Vice Secretary",
+    },
+    {
+        img: Priyanshu_shri,
+        name: "Priyanshi Srivastava",
+        post: "Core Member",
+    },
+    {
+        img: Shreyansh_Kushwaha,
+        name: "Shreyansh Kushwaha",
+        post: "Core Member",
+    },
+    {
+        img: Venkata_surya,
+        name: "Venkata Surya Sundar Vadali",
+        post: "Core Member",
+    },
+    {
+        img: Vishal_Sharma,
+        name: "Vishal Sharma",
+        post: "Core Member",
+    },
+];
+export const bis: CrewProps[] = []
+export const tsoc: CrewProps[] = [
+    {
+        img: Shivaibhav,
+        name: "Shivaibhav Dewangan",
+        post: "Club Secretary",
+    },
+    {
+        img: Gaurish,
+        name: "Gaurish Ojha",
+        post: "Vice Secretary",
+    },
+    {
+        img: Dhyanendra,
+        name: "Dhyaanendra Tripathi",
+        post: "Core Member",
+    },
+    {
+        img: Swasti_Srivastava,
+        name: "Swasti Srivastava",
+        post: "Core Member",
+    },
+    {
+        img: Manav_Raj,
+        name: "Manav Raj Singh",
+        post: "Core Member",
+    },
+    {
+        img: Ashutosh_Jaiswal,
+        name: "Ashutosh Jaiswal",
+        post: "Core Member",
+    }
+];
+
+export const aiml: CrewProps[] = [
+    {
+        img: YashV_Gautum,
+        name: "Yash Vardhan Gautam",
+        post: "Club Secretary",
+    },
+    {
+        img: Sanskar_Singh,
+        name: "Sanskar Singh",
+        post: "Vice Secretary",
+    },
+    {
+        img: Shaurya_Malhan,
+        name: "Shaurya Malhan",
+        post: "Core Member",
+    },
+    {
+        img: Arya_Mahesh,
+        name: "Arya Mahesh Bhiwapurkar",
+        post: "Core Member",
+    },
+    {
+        img: Vaibhav_Suntwal,
+        name: "Vaibhav Suntwal",
+        post: "Core Member",
+    },
+    {
+        img: Swastika_Satya,
+        name: "Swastika Satya",
+        post: "Core Member",
+    },
+];
