+++ features/club/technical/BasicMenu.tsx
@@ -151,4 +151,6 @@ const clubs = [
     color: "#BB9AF7",
   },
   { name: "// Comet", url: "/clubs/technical/comet", color: "#FF9E64" },
+  { name: "// BIS", url: "/clubs/technical/bis", color: "#CA50F5" },
+  { name: "// IIC", url: "/clubs/technical/iic", color: "#FFDE31" },
 ];
