+++ pages/clubs/media/design.tsx
@@ -8,15 +8,7 @@ import DesignBg from "../../../assets/media_cell_bg/design.jpg";
 import DesignLogo from "../../../assets/media_cell_bg/Media.png";
 import Crew from "../../../features/club/common/Crew";
 import { StaticImageData } from "next/image";
-
-// import crew
-
-import A_Satya_Kartik from "../../../assets/team_23-24/Male-member.jpeg";
-import S_Chandu from "../../../assets/team_23-24/Male-member.jpeg";
-import Ojash_Verma from "../../../assets/team_23-24/Male-member.jpeg";
-import Tanishka from "../../../assets/team_23-24/Female-member.jpeg";
-import Trisul from "../../../assets/team_23-24/Trisul.jpeg";
-import Shreyansh_Kushwaha from "../../../assets/team_23-24/Shreyansh_Kushwaha.jpg";
+import { design } from "../../../data/MediaClubCrew";
 
 interface DesignProps {}
 
@@ -55,7 +47,7 @@ const Design: FunctionComponent<DesignProps> = () => {
           team, documenting every posts, blogs or captions, thus suopporting the
           entire SAC.
         </ClubDescription>
-        <Crew props={crew} />
+        <Crew props={design} />
         <br />
       </ClubsLayout>
     </>
@@ -64,47 +56,8 @@ const Design: FunctionComponent<DesignProps> = () => {
 
 export default Design;
 
-interface CrewProps {
-  img: StaticImageData;
-  name: string;
-  post: string;
-}
-
 interface GalleryProps {
   name: string;
   img: StaticImageData[];
   date: string;
 }
-
-const crew: CrewProps[] = [
-  {
-    img: A_Satya_Kartik,
-    name: "A Satya Karthik",
-    post: "Head",
-  },
-  {
-    img: S_Chandu,
-    name: "S.Chandu",
-    post: "Core Member",
-  },
-  {
-    img: Ojash_Verma,
-    name: "Ojash Verma",
-    post: "Core Member",
-  },
-  {
-    img: Tanishka,
-    name: "Tanishka Chauhan",
-    post: "Core Member",
-  },
-  {
-    img: Shreyansh_Kushwaha,
-    name: "Shreyansh Kushwaha",
-    post: "Core Member",
-  },
-  {
-    img: Trisul,
-    name: "Trishul",
-    post: "Core Member",
-  },
-];
