+++ pages/clubs/technical/comet.tsx
@@ -10,14 +10,7 @@ import cometBgTop from "../../../assets/technical_clubs_bg/bg_comet_top.svg";
 import cometBgBottom from "../../../assets/technical_clubs_bg/bg_comet_bottom.svg";
 import cometLogo from "../../../assets/technical_clubs_bg/comet_logo.png";
 import { StaticImageData } from "next/image";
-
-// import crew
-import Ishita_Tegar from "../../../assets/team_23-24/Ishita_Tegar.jpg";
-import Umesh_Sinha from "../../../assets/team_23-24/Umesh_Sinha.jpg";
-import T_Srivedanarayana from "../../../assets/team_23-24/Male-member.jpeg";
-import Alankar_Saxena from "../../../assets/team_23-24/Alankar_Saxena.jpg";
-import Namita from "../../../assets/team_23-24/Namita_shah.jpg";
-import Krishna_Aggarwal from "../../../assets/team_23-24/Krishna_Agrawal.jpg";
+import { comet } from "../../../data/TechnicalClubCrew";
 
 // import events
 import arduino1 from "../../../assets/club_event_pics/technical/arduino1.jpeg";
@@ -42,7 +35,7 @@ const Comet = () => {
           event2Date="April 10, 2023"
           event2Desc="Workshop on R-pi"
         />
-        <ClubDescription bg={cometBg} >
+        <ClubDescription bg={cometBg}>
           The club was established in 2018 under the name CECoT (Community of
           electronics and communication techies) and was later renamed Community
           of Electronics. The club was founded by Harsh Singh, Akhil Kumar,
@@ -63,7 +56,7 @@ const Comet = () => {
             pb: 6,
           }}
         >
-          <Crew props={crew} />
+          <Crew props={comet} />
           <Gallery props={gallery} />
         </Box>
       </ClubsLayout>
@@ -73,45 +66,11 @@ const Comet = () => {
 
 export default Comet;
 
-interface CrewProps {
-  img: StaticImageData;
-  name: string;
-  post: string;
-}
-
 interface GalleryProps {
   name: string;
   img: StaticImageData[];
   date: string;
 }
-
-const crew: CrewProps[] = [
-  {
-    img: Ishita_Tegar,
-    name: "Ishita Tegar",
-    post: "Club Secretary",
-  },
-  {
-    img: Umesh_Sinha,
-    name: "Umesh Sinha",
-    post: "Vice Secretary",
-  },
-  {
-    img: Alankar_Saxena,
-    name: "Alankar Saxena",
-    post: "Core Member",
-  },
-  {
-    img: Namita,
-    name: "Namita Shah",
-    post: "Core Member",
-  },
-  {
-    img: Krishna_Aggarwal,
-    name: "Krishna Aggarwal",
-    post: "Core Member",
-  },
-];
 
 const gallery: GalleryProps[] = [
   {
