+++ pages/clubs/technical/iic.tsx
@@ -10,14 +10,7 @@ import IICBgTop from "../../../assets/technical_clubs_bg/bg_IIC_top.png";
 import IICBgBottom from "../../../assets/technical_clubs_bg/bg_IIC_bottom.png";
 import IICLogo from "../../../assets/technical_clubs_bg/IIC_logo.png";
 import { StaticImageData } from "next/image";
-
-// import crew
-import Shivaibhav from "../../../assets/team_23-24/Shivaibhav_Dewangan.jpg";
-import Gaurish from "../../../assets/team_23-24/Gaurish_Ojha.jpg";
-import Dhyanendra from "../../../assets/team_23-24/Dhyanendra_Tripathi.jpg";
-import Swasti_Srivastava from "../../../assets/team_23-24/Swasti_Srivastava.jpeg";
-import Manav_Raj from "../../../assets/team_23-24/Manav_Raj_Singh.jpg";
-import Ashutosh_Jaiswal from "../../../assets/team_23-24/Ashutosh_Jaiswal.jpeg";
+import { iic } from "../../../data/TechnicalClubCrew";
 
 // import events
 import bit1 from "../../../assets/club_event_pics/technical/bit1.jpeg";
@@ -64,7 +57,7 @@ const IIC: FunctionComponent<IICProps> = () => {
             pb: 6,
           }}
         >
-          {/* <Crew props={crew} />
+          {/* <Crew props={iib} />
           <Gallery props={gallery} /> */}
         </Box>
       </ClubsLayout>
@@ -74,70 +67,12 @@ const IIC: FunctionComponent<IICProps> = () => {
 
 export default IIC;
 
-interface CrewProps {
-  img: StaticImageData;
-  name: string;
-  post: string;
-}
-
 interface GalleryProps {
   name: string;
   img: StaticImageData[];
   date: string;
 }
 
-const crew: CrewProps[] = [
-  {
-    img: Shivaibhav,
-    name: "Shivaibhav Dewangan",
-    post: "Club Secretary",
-  },
-  {
-    img: Gaurish,
-    name: "Gaurish Ojha",
-    post: "Vice Secretary",
-  },
-  {
-    img: Dhyanendra,
-    name: "Dhyaanendra Tripathi",
-    post: "Core Member",
-  },
-  {
-    img: Swasti_Srivastava,
-    name: "Swasti Srivastava",
-    post: "Core Member",
-  },
-  {
-    img: Manav_Raj,
-    name: "Manav Raj Singh",
-    post: "Core Member",
-  },
-  {
-    img: Ashutosh_Jaiswal,
-    name: "Ashutosh Jaiswal",
-    post: "Core Member",
-  },
-  {
-    img: Dhyanendra,
-    name: "Dhyaanendra Tripathi",
-    post: "Core Member",
-  },
-  {
-    img: Swasti_Srivastava,
-    name: "Swasti Srivastava",
-    post: "Core Member",
-  },
-  {
-    img: Manav_Raj,
-    name: "Manav Raj Singh",
-    post: "Core Member",
-  },
-  {
-    img: Ashutosh_Jaiswal,
-    name: "Ashutosh Jaiswal",
-    post: "Core Member",
-  },
-];
 const gallery: GalleryProps[] = [
   {
     name: "Bit-2-Byte",
