+++ pages/clubs/technical/inquizitive.tsx
@@ -10,14 +10,7 @@ import inquizitiveBgTop from "../../../assets/technical_clubs_bg/bg_inquizitive_
 import inquizitiveBgBottom from "../../../assets/technical_clubs_bg/bg_inquizitive_bottom.svg";
 import inquizitiveLogo from "../../../assets/technical_clubs_bg/inquizitive_logo.jpg";
 import { StaticImageData } from "next/image";
-
-// import crew
-import Ayush_Yadav from "../../../assets/team_23-24/Ayush_Yadav.jpg";
-import Bhavik from "../../../assets/team_23-24/Male-member.jpeg";
-import Aniket_Pandey from "../../../assets/team_23-24/Aniket_Pandey.jpg";
-import Arman_Singh_Kshatri from "../../../assets/team_23-24/Male-member.jpeg";
-import Harsh_Shrivastava from "../../../assets/team_23-24/Harsh_Shrivastava.jpg";
-import Lokesh_Harmani from "../../../assets/team_23-24/Male-member.jpeg";
+import { inquizitive } from "../../../data/TechnicalClubCrew";
 
 // import events
 import quiz1 from "../../../assets/club_event_pics/technical/quiz1.jpeg";
@@ -63,7 +56,7 @@ const Inquizitive = () => {
             pb: 6,
           }}
         >
-          <Crew props={crew} />
+          <Crew props={inquizitive} />
           <Gallery props={gallery} />
         </Box>
       </ClubsLayout>
@@ -73,50 +66,11 @@ const Inquizitive = () => {
 
 export default Inquizitive;
 
-interface CrewProps {
-  img: StaticImageData;
-  name: string;
-  post: string;
-}
-
 interface GalleryProps {
   name: string;
   img: StaticImageData[];
   date: string;
 }
-
-const crew: CrewProps[] = [
-  {
-    img: Ayush_Yadav,
-    name: "Ayush Yadav",
-    post: "Club Secretary",
-  },
-  {
-    img: Bhavik,
-    name: "Bhavik Sachdev",
-    post: "Vice Secretary",
-  },
-  {
-    img: Aniket_Pandey,
-    name: "Aniket Pandey",
-    post: "Core Member",
-  },
-  {
-    img: Arman_Singh_Kshatri,
-    name: "Arman Singh Kshatri",
-    post: "Core Member",
-  },
-  {
-    img: Harsh_Shrivastava,
-    name: "Harsh Shrivastava",
-    post: "Core Member",
-  },
-  {
-    img: Lokesh_Harmani,
-    name: "Lokesh Hariramani",
-    post: "Core Member",
-  },
-];
 
 const gallery: GalleryProps[] = [
   {
